Object 조작 함수 모듈화
