+++ service/PerfumeService.js
@@ -14,15 +14,12 @@ const {
     FailedToCreateError,
 } = require('../utils/errors/errors.js');
 
-function updateRows(result, ...jobs) {
-    const perfumeList = result.rows;
-    result.rows = perfumeList.map((it) => {
-        return jobs.reduce((prev, cur) => {
-            return cur(prev);
-        }, it);
-    });
-    return result;
-}
+const {
+    updateRows,
+    removeKeyJob,
+    extractJob,
+    flatJob,
+} = require('../utils/func.js');
 
 function isLikeJob(likePerfumeList) {
     const likeMap = likePerfumeList.reduce((prev, cur) => {
@@ -32,38 +29,6 @@ function isLikeJob(likePerfumeList) {
     return (obj) => {
         const ret = Object.assign({}, obj);
         ret.isLiked = likeMap[obj.perfumeIdx] ? true : false;
-        return ret;
-    };
-}
-
-function removeKeyJob(...keys) {
-    return (obj) => {
-        const ret = Object.assign({}, obj);
-        keys.forEach((it) => {
-            delete ret[it];
-        });
-        return ret;
-    };
-}
-
-function extractJob(key, ...fields) {
-    return (obj) => {
-        const ret = Object.assign({}, obj);
-        fields.forEach((it) => {
-            ret[it[1]] = obj[key][it[0]];
-        });
-        delete ret[key];
-        return ret;
-    };
-}
-
-function flatJob(...keys) {
-    return (obj) => {
-        let ret = Object.assign({}, obj);
-        keys.forEach((key) => {
-            ret = Object.assign(ret, obj[key]);
-            delete ret[key];
-        });
         return ret;
     };
 }
